+++ Model/Util.php
@@ -2,33 +2,43 @@
 
 namespace FlowCommerce\FlowConnector\Model;
 
-use \FlowCommerce\FlowConnector\Exception\FlowException;
+use Exception;
+use FlowCommerce\FlowConnector\Exception\FlowException;
+use FlowCommerce\FlowConnector\Model\Api\Auth;
+use FlowCommerce\FlowConnector\Model\Api\UrlBuilder;
 use FlowCommerce\FlowConnector\Model\GuzzleHttp\Client as GuzzleClient;
-use \Magento\Store\Model\ScopeInterface;
-use \Magento\Framework\App\Config\ScopeConfigInterface as ScopeConfig;
-use \Magento\Framework\Module\ModuleListInterface as ModuleList;
-use \Magento\Store\Model\StoreManagerInterface as StoreManager;
-use \Psr\Log\LoggerInterface as Logger;
-use \Zend\Http\ClientFactory as HttpClientFactory;
-use \Zend\Http\Request;
-use \Zend\Http\Client\Adapter\Exception\RuntimeException;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Store\Api\Data\StoreInterface;
+use Magento\Store\Model\ScopeInterface;
+use Magento\Framework\App\Config\ScopeConfigInterface as ScopeConfig;
+use Magento\Framework\Module\ModuleListInterface as ModuleList;
+use Magento\Store\Model\StoreManagerInterface as StoreManager;
+use Psr\Log\LoggerInterface as Logger;
+use Zend\Http\Client;
+use Zend\Http\ClientFactory as HttpClientFactory;
+use Zend\Http\Request;
+use Zend\Http\Client\Adapter\Exception\RuntimeException;
+use Zend\Http\Response;
 
 /**
  * Utility class for Flow settings and endpoints.
  */
-class Util {
-
+class Util
+{
     // Store configuration key for Flow Enabled
     const FLOW_ENABLED = 'flowcommerce/flowconnector/enabled';
 
-    // Store configuration key for Flow Organization Id
-    const FLOW_ORGANIZATION_ID = 'flowcommerce/flowconnector/organization_id';
-
-    // Store configuration key for Flow API Token
-    const FLOW_API_TOKEN = 'flowcommerce/flowconnector/api_token';
-
     // Store configuration key for Flow Invoice Event
     const FLOW_INVOICE_EVENT = 'flowcommerce/flowconnector/invoice_event';
+
+    // Store configuration key for Flow Shipment Event
+    const FLOW_SHIPMENT_EVENT = 'flowcommerce/flowconnector/shipment_event';
+
+    // Store configuration key for Flow Invoice Event
+    const FLOW_INVOICE_SEND_EMAIL = 'flowcommerce/flowconnector/invoice_email';
+
+    // Store configuration key for Flow Invoice Event
+    const FLOW_SHIPMENT_SEND_EMAIL = 'flowcommerce/flowconnector/shipment_email';
 
     // Flow API base endpoint
     const FLOW_API_BASE_ENDPOINT = 'https://api.flow.io/';
@@ -49,143 +59,131 @@ class Util {
     const HTTP_USERAGENT = 'Flow-M2';
 
     /**
-     * @var Logger
+     * @var Auth
      */
-    protected $logger;
-
-    /**
-     * @var ScopeConfig
-     */
-    protected $scopeConfig;
-
-    /**
-     * @var StoreManager
-     */
-    protected $storeManager;
-
-    /**
-     * @var ModuleList
-     */
-    protected $moduleList;
-
-    /**
-     * @var string
-     */
-    protected $moduleVersion;
+    private $auth;
 
     /**
      * @var HttpClientFactory
      */
-    protected $httpClientFactory;
+    private $httpClientFactory;
+
+    /**
+     * @var Logger
+     */
+    private $logger;
+
+    /**
+     * @var ScopeConfig
+     */
+    private $scopeConfig;
+
+    /**
+     * @var StoreManager
+     */
+    private $storeManager;
+
+    /**
+     * @var ModuleList
+     */
+    private $moduleList;
+
+    /**
+     * @var UrlBuilder
+     */
+    private $urlBuilder;
 
     /**
      * Util constructor.
+     * @param Auth $auth
      * @param Logger $logger
+     * @param HttpClientFactory $httpClientFactory
+     * @param ModuleList $moduleList
      * @param ScopeConfig $scopeConfig
      * @param StoreManager $storeManager
-     * @param ModuleList $moduleList
-     * @param HttpClientFactory $httpClientFactory
+     * @param UrlBuilder $urlBuilder
      */
     public function __construct(
+        Auth $auth,
+        HttpClientFactory $httpClientFactory,
         Logger $logger,
+        ModuleList $moduleList,
         ScopeConfig $scopeConfig,
         StoreManager $storeManager,
-        ModuleList $moduleList,
-        HttpClientFactory $httpClientFactory
+        UrlBuilder $urlBuilder
     ) {
+        $this->auth = $auth;
+        $this->httpClientFactory = $httpClientFactory;
         $this->logger = $logger;
+        $this->moduleList = $moduleList;
         $this->scopeConfig = $scopeConfig;
         $this->storeManager = $storeManager;
-        $this->moduleList = $moduleList;
-        $this->moduleVersion = $this->moduleList->getOne('FlowCommerce_FlowConnector')['setup_version'];
-        $this->httpClientFactory = $httpClientFactory;
+        $this->urlBuilder = $urlBuilder;
     }
 
     /**
-    * Set the logger (used by console command).
-    */
-    public function setLogger($logger) {
+     * Set the logger (used by console command).
+     * @param Logger $logger
+     * @return void
+     */
+    public function setLogger($logger)
+    {
         $this->logger = $logger;
     }
 
     /**
      * Returns true if Flow is enabled in the Admin Store Configuration.
-     * @param storeId ID of store, if null defaults to current store.
+     * @param int|null $storeId
+     * @return bool
+     * @throws NoSuchEntityException
      */
-    public function isFlowEnabled($storeId = null) {
-        if (is_null($storeId)) {
+    public function isFlowEnabled($storeId = null)
+    {
+        if ($storeId === null) {
             $storeId = $this->getCurrentStoreId();
         }
-
-        return $this->scopeConfig->getValue(self::FLOW_ENABLED, ScopeInterface::SCOPE_STORE, $storeId);
+        return (bool) $this->scopeConfig->getValue(self::FLOW_ENABLED, ScopeInterface::SCOPE_STORE, $storeId);
     }
 
     /**
-     * Returns the Flow Organization Id set in the Admin Store Configuration.
-     * @param storeId ID of store, if null defaults to current store.
+     * Returns the Flow Client user agent
+     * @return string
      */
-    public function getFlowOrganizationId($storeId = null) {
-        if (is_null($storeId)) {
-            $storeId = $this->getCurrentStoreId();
-        }
-
-        return $this->scopeConfig->getValue(self::FLOW_ORGANIZATION_ID, ScopeInterface::SCOPE_STORE, $storeId);
-    }
-
-    /**
-     * Returns the Flow API Token set in the Admin Store Configuration.
-     * @param storeId ID of store, if null defaults to current store.
-     */
-    public function getFlowApiToken($storeId = null) {
-        if (is_null($storeId)) {
-            $storeId = $this->getCurrentStoreId();
-        }
-
-        return $this->scopeConfig->getValue(self::FLOW_API_TOKEN, ScopeInterface::SCOPE_STORE, $storeId);
-    }
-
-    /**
-     * Returns the Flow API endpoint with the specified url stub.
-     * @param urlStub Url stub for the client
-     * @param storeId ID of store, if null defaults to current store.
-     */
-    public function getFlowApiEndpoint($urlStub, $storeId = null) {
-        if (is_null($storeId)) {
-            $storeId = $this->getCurrentStoreId();
-        }
-
-        return self::FLOW_API_BASE_ENDPOINT .
-            $this->getFlowOrganizationId($storeId) . $urlStub;
-    }
-
-    /**
-     * Returns the Flow Client user agent.
-     */
-    public function getFlowClientUserAgent() {
-        return GuzzleClient::HTTP_USERAGENT . '-' . $this->moduleVersion;
+    public function getFlowClientUserAgent()
+    {
+        return GuzzleClient::HTTP_USERAGENT . '-' . $this->getModuleVersion();
     }
 
     /**
      * Returns a Zend Client preconfigured for the Flow API.
-     * @param urlStub Url stub for the client
-     * @param storeId ID of store, if null defaults to current store.
+     * @param string $urlStub
+     * @param int|null $storeId
+     * @return Client
+     * @throws NoSuchEntityException
      */
-    public function getFlowClient($urlStub, $storeId = null) {
-        if (is_null($storeId)) {
+    public function getFlowClient($urlStub, $storeId = null)
+    {
+        if ($storeId === null) {
             $storeId = $this->getCurrentStoreId();
         }
 
-        $useragent = $this->getFlowClientUserAgent();
-        $url = $this->getFlowApiEndpoint($urlStub, $storeId);
-        $this->logger->info('Flow Client [' . $useragent . '] URL: ' . $url);
+        $userAgent = $this->getFlowClientUserAgent();
+        $url = $this->urlBuilder->getFlowApiEndpoint($urlStub, $storeId);
+        $this->logger->info('Flow Client [' . $userAgent . '] URL: ' . $url);
 
-        $client = $this->getHttpClient($url, $useragent);
+        $client = $this->getHttpClient($url, $userAgent);
         $client->setMethod(Request::METHOD_GET);
-        $client->setAuth($this->getFlowApiToken($storeId), '');
+        $client->setAuth($this->auth->getFlowApiToken($storeId), '');
         $client->setEncType('application/json');
         return $client;
     }
 
+    /**
+     * Returns Zend client instance
+     * @param $url
+     * @param $userAgent
+     * @return Client
+     */
     private function getHttpClient($url, $userAgent)
     {
         return $this->httpClientFactory->create([
@@ -198,18 +196,30 @@ class Util {
     }
 
     /**
-     * Wrapper function to retry on timeout for http client send().
-     * @param client Flow http client
-     * @param numRetries Defaults to 3 retries
+     * Returns current Flow Connector version
+     * @return string
      */
-    public function sendFlowClient($client, $numRetries = 3) {
+    private function getModuleVersion()
+    {
+        return (string) $this->moduleList->getOne('FlowCommerce_FlowConnector')['setup_version'];
+    }
+
+    /**
+     * Wrapper function to retry on timeout for http client send().
+     * @param Client $client
+     * @param int|null $numRetries
+     * @return Response
+     */
+    public function sendFlowClient($client, $numRetries = 3)
+    {
         try {
             return $client->send();
         } catch (RuntimeException $e) {
             if ($numRetries <= 0) {
                 throw $e;
             } else {
-                $this->logger->info('Error sending client request, retries remaining: ' . $numRetries . ', trying again in ' . self::FLOW_CLIENT_RETRY_DELAY . ' seconds');
+                $this->logger->info('Error sending client request, retries remaining: ' . $numRetries .
+                    ', trying again in ' . self::FLOW_CLIENT_RETRY_DELAY . ' seconds');
                 sleep(self::FLOW_CLIENT_RETRY_DELAY);
                 return $this->sendFlowClient($client, $numRetries - 1);
             }
@@ -218,38 +228,43 @@ class Util {
 
     /**
      * Returns the Flow Checkout Url
-     * @param storeId ID of store, if null defaults to current store.
+     * @param int|null $storeId
+     * @throws NoSuchEntityException
+     * @return string
      */
-    public function getFlowCheckoutUrl($storeId = null) {
-        if (is_null($storeId)) {
+    public function getFlowCheckoutUrl($storeId = null)
+    {
+        if ($storeId === null) {
             $storeId = $this->getCurrentStoreId();
         }
 
         return self::FLOW_CHECKOUT_BASE_URL .
-            $this->getFlowOrganizationId($storeId) . '/order/';
+            $this->auth->getFlowOrganizationId($storeId) . '/order/';
     }
 
     /**
      * Returns the ID of the current store
      * @return int
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @throws NoSuchEntityException
      */
-    public function getCurrentStoreId() {
+    public function getCurrentStoreId()
+    {
         return $this->storeManager->getStore()->getId();
     }
 
     /**
      * Notifies Flow cross dock that order is enroute.
-     *
      * https://docs.flow.io/module/logistics/resource/shipping_notifications#put-organization-shipping-notifications-key
      * https://docs.flow.io/type/shipping-label-package
-     *
-     * @param order The Magento order object.
-     * @param trackingNumber The tracking number for order sent to cross dock.
-     * @param shippingPackageLabel A Flow Shipping Label Package object.
-     * @param service Carrier service level used for generation and shipment of this label.
+     * @param $order - The Magento order object.
+     * @param $trackingNumber - The tracking number for order sent to cross dock.
+     * @param $shippingLabelPackage - A Flow Shipping Label Package object.
+     * @param $service - Carrier service level used for generation and shipment of this label.
+     * @throws Exception
+     * @TODO
      */
-    public function notifyCrossDock($order, $trackingNumber, $shippingLabelPackage, $service) {
+    public function notifyCrossDock($order, $trackingNumber, $shippingLabelPackage, $service)
+    {
         $flowOrder = $this->flowOrderFactory->create()->find('order_id', $order->getId());
         $storeId = $order->getStoreId();
 
@@ -273,17 +288,21 @@ class Util {
             $this->logger->error('Notify Cross Dock: failed');
             $this->logger->error('Status code: ' . $response->getStatusCode());
             $this->logger->error('Body: ' . $response->getBody());
-            throw new FlowException('Failed to notify cross dock with tracking number ' . $trackingNumber . ': ' . $response->getBody());
+            throw new FlowException('Failed to notify cross dock with tracking number ' . $trackingNumber .
+                ': ' . $response->getBody());
         }
     }
 
     /**
      * Returns an array of Stores that are enabled for Flow.
+     * @return StoreInterface[]
+     * @throws NoSuchEntityException
      */
-    public function getEnabledStores() {
+    public function getEnabledStores()
+    {
         $stores = [];
         foreach ($this->storeManager->getStores() as $store) {
-            if ($this->isFlowEnabled($store->getStoreId())) {
+            if ($this->isFlowEnabled($store->getId())) {
                 array_push($stores, $store);
             }
         }
@@ -292,17 +311,70 @@ class Util {
 
     /**
      * Returns Flow Invoice Event
-     *
      * @see \FlowCommerce\FlowConnector\Model\Config\Source\InvoiceEvent::toOptionArray()
-     *
      * @param null $storeId
      * @return int
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @throws NoSuchEntityException
      */
-    public function getFlowInvoiceEvent($storeId = null) {
+    public function getFlowInvoiceEvent($storeId = null)
+    {
         if ($storeId === null) {
             $storeId = $this->getCurrentStoreId();
         }
         return (int)$this->scopeConfig->getValue(self::FLOW_INVOICE_EVENT, ScopeInterface::SCOPE_STORE, $storeId);
+    }
+
+    /**
+     * Returns Flow Shipment Event
+     *
+     * @see \FlowCommerce\FlowConnector\Model\Config\Source\ShipmentEvent::toOptionArray()
+     *
+     * @param int|null $storeId
+     * @return int
+     * @throws NoSuchEntityException
+     */
+    public function getFlowShipmentEvent($storeId = null)
+    {
+        if ($storeId === null) {
+            $storeId = $this->getCurrentStoreId();
+        }
+        return (int)$this->scopeConfig->getValue(self::FLOW_SHIPMENT_EVENT, ScopeInterface::SCOPE_STORE, $storeId);
+    }
+
+    /**
+     * Returns true if send invoice email is enabled in the Admin Store Configuration.
+     * @param int|null $storeId
+     * @return bool
+     * @throws NoSuchEntityException
+     */
+    public function sendInvoiceEmail($storeId = null)
+    {
+        if ($storeId === null) {
+            $storeId = $this->getCurrentStoreId();
+        }
+        return (bool) $this->scopeConfig->getValue(
+            self::FLOW_INVOICE_SEND_EMAIL,
+            ScopeInterface::SCOPE_STORE,
+            $storeId
+        );
+    }
+
+    /**
+     * Returns true if send shipment email is enabled in the Admin Store Configuration.
+     * @param int|null $storeId
+     * @return bool
+     * @throws NoSuchEntityException
+     */
+    public function sendShipmentEmail($storeId = null)
+    {
+        if ($storeId === null) {
+            $storeId = $this->getCurrentStoreId();
+        }
+
+        return (bool) $this->scopeConfig->getValue(
+            self::FLOW_SHIPMENT_SEND_EMAIL,
+            ScopeInterface::SCOPE_STORE,
+            $storeId
+        );
     }
 }